+++ ydb/public/sdk/cpp/tests/unit/client/coordination/coordination_ut.cpp
@@ -102,7 +102,7 @@ Y_UNIT_TEST_SUITE(Coordination) {
         UNIT_ASSERT_VALUES_EQUAL_C(res.GetStatus(), EStatus::SUCCESS, res.GetIssues().ToString());
 
         auto session = res.ExtractResult();
-        stoppedFuture.Wait();
+        UNIT_ASSERT(stoppedFuture.Wait(TDuration::Seconds(10)));
         auto endTimestamp = TInstant::Now();
         auto elapsed = endTimestamp - startTimestamp;
 
@@ -113,7 +113,7 @@ Y_UNIT_TEST_SUITE(Coordination) {
         UNIT_ASSERT_VALUES_EQUAL_C(res2.GetStatus(), EStatus::TIMEOUT, res2.GetIssues().ToString());
     }
 
-    Y_UNIT_TEST(SessionCancelByDriver) {
+    Y_UNIT_TEST(DriverStopPreservesSessionAfterHandlesAreReleased) {
         TPortManager pm;
 
         TMockCoordinationService coordinationService;
@@ -150,26 +150,30 @@ Y_UNIT_TEST_SUITE(Coordination) {
             .OnStopped([stoppedPromise]() mutable {
                 stoppedPromise.SetValue();
             })
-            .Timeout(TDuration::MilliSeconds(1000));
+            .Timeout(TDuration::Seconds(30));
 
-        auto startTimestamp = TInstant::Now();
-        auto res = client->StartSession("/Some/Path", settings).ExtractValueSync();
+        auto started = client->StartSession("/Some/Path", settings);
+        UNIT_ASSERT(started.Wait(TDuration::Seconds(10)));
+        auto res = started.ExtractValue();
         UNIT_ASSERT_VALUES_EQUAL_C(res.GetStatus(), EStatus::SUCCESS, res.GetIssues().ToString());
 
         auto session = res.ExtractResult();
 
         client.reset();
-        driver->Stop();
+        driver->Stop(true);
         driver.reset();
+        UNIT_ASSERT(!stoppedFuture.IsReady());
 
-        stoppedFuture.Wait();
-        auto endTimestamp = TInstant::Now();
-        auto elapsed = endTimestamp - startTimestamp;
+        auto ping = session.Ping();
+        UNIT_ASSERT(ping.Wait(TDuration::Seconds(10)));
+        auto pingResult = ping.ExtractValue();
+        UNIT_ASSERT_VALUES_EQUAL_C(pingResult.GetStatus(), EStatus::SUCCESS, pingResult.GetIssues().ToString());
 
-        UNIT_ASSERT_C(elapsed < TDuration::Seconds(1), "Elapsed time too large: " << elapsed);
-
-        auto res2 = session.Close().ExtractValueSync();
-        UNIT_ASSERT_VALUES_EQUAL_C(res2.GetStatus(), EStatus::CLIENT_CANCELLED, res2.GetIssues().ToString());
+        auto closed = session.Close();
+        UNIT_ASSERT(closed.Wait(TDuration::Seconds(10)));
+        auto closeResult = closed.ExtractValue();
+        UNIT_ASSERT_VALUES_EQUAL_C(closeResult.GetStatus(), EStatus::SUCCESS, closeResult.GetIssues().ToString());
+        UNIT_ASSERT(stoppedFuture.Wait(TDuration::Seconds(10)));
     }
 
     Y_UNIT_TEST(SessionDropsDriverFromStateCallback) {